de-duplicate code between ASTNeuron and ASTNeuronOrSynapse
